Refactor models and imports: move Post model to api.models, update import paths in services and graphs
