+++ backend/src/api/services/storage_service.py
@@ -2,7 +2,7 @@ import uuid
 import random
 from typing import List, Dict
 from ..core.logging import logger
-from ...models import Post
+from ..models.models import Post
 
 class StorageService:
     """
