+++ backend/src/graphs/workflow.py
@@ -1,6 +1,6 @@
 from langgraph.graph import StateGraph, END
 from ..core.logging import logger
-from .state import ContentState
+from ..api.models.models import ContentState
 from .nodes import generate_node, validate_node, error_node, router
 
 async def generate_content_workflow(prompt: str) -> str:
